Update default statusline to remove total_time and put vol at right
This is a more intuitive place for volume, particularly following:
* 971204b - Add mouse bindings for right third of status bar

The %{total} time always seemed out of place to me and now that we have
added it to the header of the Sorted view it is even less useful here.

This `format_statusline` will be the default for new users, but existing
users will not see the change.